chore: page service includes ChatStream functionality rather than relying on an Answer Service
